Clarify message when kernel config is out of date
When importing the user provided kernel config, we test that it
works by running "make oldconfig" with a timeout. If the timeout
occurs, it means that the configurtion is not complete and the
build exits with an error. Up to now, the error message was a
cryptic "kernel configuration is not up to date". Change the
message so that the user understands what is going on and how to
fix the issue.

Reported on
https://github.com/automate-lfs/jhalfs/issues/23
Fixes this issue