Make test-mismatch configurable
Presently, it is not possible to disable testing a mismatch
between an extracted directory name and a package version. This
may be a problem in some cases (for example if using base+patch
for the kernel). The infrastructure is already there. We just
need to add the possibility to set TEST_MISMATCH.
This commit adds it to Config.in