Don't download any <ulink> in chapter 3
Recently, there was a note added to the linux kernel informations,
explaining how to use a base tarball + patch to extract the build
tree of the kernel. This note uses <ulink> tags in examples. But
our "urls.xsl" stylesheet considers any <ulink> descendant of a
<varlistentry> is a file to be downloaded, and should be followed
by a valid mds sum. The lack of valid md5 sum is what triggers
an error, actually.
We fix this by assuming that any varlistentry/listitem/para/ulink
is a valid url to download (that is, we refine the path search).

Fix error reported by Marcin Dulak in
https://lists.linuxfromscratch.org/sympa/arc/lfs-support/2025-09/msg00018.html
and
https://lists.linuxfromscratch.org/sympa/arc/alfs-discuss/2025-09/msg00002.html