Make "devices" an "order only" target
In the generated Makefile, we have a number of targets that depend
on the "devices" target, because virtual kernel fs have to be
mounted when running those targets. Problem is that "devices" is
.PHONY, so that it is unconditionally run, and then, all targets
that depend on it are run to. If restarting "make" after say, a
failure, this means some targets may be run again. Usually, it
is not a problem, because those targets only call make with another
target, and if this other target has been run, then nothing is done.
But it may happen that a target depending on "devices" also does some
file creation/removal or some kind of garbage collection (remove lfs
user e.g.). Trying to remove again a removed file, for example, would
fail. Removing the lfs user produces an error (ignored because we
add a dash before the command), etc.

Instead of guarding (ignoring) against errors, it is much better to
make "devices" an "order only" target (add a | char before it).