Make runtime dependencies optional "before" dependencies
The problem is that some runtime dependencies may be used in tests.
To have them built for tests (which means including optional
dependencies), move an "after" dependency to "before" "optional".
This way, the runtime dependency will be built before if
possible. Note that we don't move them to optional if they
depend on the package that declares them as runtime dependencies.
THe cycle breaker may also not consider them if it creates a
circular dependency...