BLFS tools: allow selecting installed packages
The program update-installed.sh opens a menu where you can select
installed packages. If a tracking file already exists, packages
that are in it are selected by default with their installed version.
They can be removed, or the version cna be changed...
If a package is not in the tracking file, it can be selected, and
by default, its version is the one in the current books. This
version can be modified too.
At the end, a new tracking file is created with what has been selected
or deselected.