Use make conditionals for BLFS and CUSTOM tools
Right now, we use a shell conditional on "ADD_BLFS_TOOLS" inside
the target mk_BLFS_TOOL, so that even if the BLFS_TOOLS are not
wanted, the target timestamp is still created. This prevents
running jhalfs again with BLFS tools enabled without erasing
this target timestamp (or the totality of what we just have built).
There is a similar problem with mk_CUSTOM_TOOL.
To fix this, we need to move the conditional to the Makefile
itself, so that the targets don't even exist if the corresponding
ADD_ variables are not "y", so that of course the timestamp files
are not created.
Bug reported by Zhang Wen on alfs-discuss